authorize

+++ app/Http/Controllers/JobController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Job;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
@@ -12,6 +13,8 @@ use Illuminate\View\View;
 
 class JobController extends Controller
 {
+
+    use AuthorizesRequests;
 
     // @desc    Show all job listings
     // @route   GET /jobs
@@ -57,7 +60,7 @@ class JobController extends Controller
             'company_website' => 'nullable|url'
         ]);
 
-        $validatedData['user_id'] = 1;
+        $validatedData['user_id'] = auth()->id();
 
         if ($request->hasFile('company_logo')) {
             $validatedData['company_logo'] = $request->file('company_logo')->store('logos', 'public');
@@ -79,6 +82,7 @@ class JobController extends Controller
     // @route   GET /jobs/{$id}/edit
     public function edit(Job $job): View
     {
+        $this->authorize('update', $job);
         return view('jobs.edit', compact('job'));
     }
 
@@ -86,6 +90,7 @@ class JobController extends Controller
     // @route   PUT /jobs/{$id}
     public function update(Request $request, Job $job): RedirectResponse
     {
+        $this->authorize('update', $job);
         $validatedData = $request->validate([
             'title' => 'required|string|max:255',
             'description' => 'required|string',
@@ -107,8 +112,6 @@ class JobController extends Controller
             'company_website' => 'nullable|url'
         ]);
 
-        $validatedData['user_id'] = 1;
-
         if ($request->hasFile('company_logo')) {
 
             Storage::delete($job->company_logo, 'logos', 'public');
@@ -125,6 +128,7 @@ class JobController extends Controller
     // @route   DELETE /jobs/{$id}
     public function destroy(Job $job): string
     {
+        $this->authorize('update', $job);
         if ($job->company_logo) {
             Storage::delete("/public/logos/$job->company_logo");
         }
